refactor: organizar cliente e fluxo de post dos links (#51)
* refactor: organize item link responsibilities

* refactor: separate item link chat and tooltip

* refactor: keep chat layout constants private
